+++ src/test/java/at/grisa/agilemetrics/producer/bitbucketserver/restclient/BitBucketServerRestClientCheckConnectionTest.java
@@ -1,6 +1,7 @@
-package at.grisa.agilemetrics.producer.sonarqube;
+package at.grisa.agilemetrics.producer.bitbucketserver.restclient;
 
 import at.grisa.agilemetrics.ApplicationConfig;
+import at.grisa.agilemetrics.producer.bitbucketserver.BitBucketServerRestClient;
 import at.grisa.agilemetrics.util.CredentialManager;
 import org.junit.Before;
 import org.junit.Rule;
@@ -13,45 +14,46 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
+import java.io.IOException;
+import java.net.URISyntaxException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+
 import static org.junit.Assert.assertTrue;
 import static org.mockserver.model.Header.header;
 import static org.mockserver.model.HttpRequest.request;
 import static org.mockserver.model.HttpResponse.response;
 
 @RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes = {SonarQubeRestClient.class, CredentialManager.class, ApplicationConfig.class})
-@TestPropertySource("classpath:sonarqube-test.properties")
-public class SonarQubeRestClientCheckConnectionTest {
+@ContextConfiguration(classes = {BitBucketServerRestClient.class, CredentialManager.class, ApplicationConfig.class})
+@TestPropertySource("classpath:bitbucket-test.properties")
+public class BitBucketServerRestClientCheckConnectionTest {
     @Autowired
-    private SonarQubeRestClient client;
+    private BitBucketServerRestClient client;
 
     @Rule
     public MockServerRule mockServerRule = new MockServerRule(this, 1080);
     private MockServerClient mockServerClient;
 
-    private boolean checkConnection;
-
     @Before
-    public void loginMockServer() {
+    public void loadPropertiesFromMockServer() throws IOException, URISyntaxException {
+        String responseBody = new String(Files.readAllBytes(Paths.get(getClass().getClassLoader().getResource("bitbucket/check.js").toURI())));
+
         mockServerClient.when(
                 request()
-                        .withMethod("POST")
-                        .withPath("/api/authentication/login")
-                        .withQueryStringParameter("login", "user")
-                        .withQueryStringParameter("password", "password")
+                        .withMethod("GET")
+                        .withPath("/rest/api/1.0/application-properties")
         )
                 .respond(
                         response()
                                 .withStatusCode(200)
                                 .withHeader(header("Content-Type", "application/json; charset=utf-8"))
+                                .withBody(responseBody)
                 );
-
-        // TODO fix console test result
-        checkConnection = true; //client.checkConnection();
     }
 
     @Test
     public void checkConnection() {
-        assertTrue("check connection", checkConnection);
+        assertTrue("check connection", client.checkConnection());
     }
 }
