+++ src/test/java/at/grisa/agilemetrics/producer/bitbucketserver/restclient/BitBucketServerRestClientCommitsTest.java
@@ -1,7 +1,8 @@
-package at.grisa.agilemetrics.producer.bitbucketserver;
+package at.grisa.agilemetrics.producer.bitbucketserver.restclient;
 
 import at.grisa.agilemetrics.ApplicationConfig;
-import at.grisa.agilemetrics.producer.bitbucketserver.restentity.Repository;
+import at.grisa.agilemetrics.producer.bitbucketserver.BitBucketServerRestClient;
+import at.grisa.agilemetrics.producer.bitbucketserver.restentity.Commit;
 import at.grisa.agilemetrics.util.CredentialManager;
 import org.junit.Before;
 import org.junit.Rule;
@@ -28,53 +29,45 @@ import static org.mockserver.model.HttpResponse.response;
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(classes = {BitBucketServerRestClient.class, CredentialManager.class, ApplicationConfig.class})
 @TestPropertySource("classpath:bitbucket-test.properties")
-public class BitBucketServerRestClientPagedEntitiesTest {
+public class BitBucketServerRestClientCommitsTest {
     @Autowired
     private BitBucketServerRestClient client;
 
     @Rule
     public MockServerRule mockServerRule = new MockServerRule(this, 1080);
     private MockServerClient mockServerClient;
-    private Collection<Repository> repositories;
+    private Collection<Commit> commits;
 
     @Before
-    public void loadReposFromMockServer() throws URISyntaxException, IOException {
-        String responseBodyPage1 = new String(Files.readAllBytes(Paths.get(getClass().getClassLoader().getResource("bitbucket/paged-entities-1.js").toURI())));
-        String responseBodyPage2 = new String(Files.readAllBytes(Paths.get(getClass().getClassLoader().getResource("bitbucket/paged-entities-2.js").toURI())));
+    public void loadCommitsFromMockServer() throws URISyntaxException, IOException {
+        String responseBody = new String(Files.readAllBytes(Paths.get(getClass().getClassLoader().getResource("bitbucket/commits.js").toURI())));
 
         mockServerClient.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/rest/api/1.0/projects/PAGED/repos")
-                        .withQueryStringParameter("start", "0")
+                        .withPath("/rest/api/1.0/projects/PRJ/repos/REPO/commits")
         )
                 .respond(
                         response()
                                 .withStatusCode(200)
                                 .withHeader(header("Content-Type", "application/json; charset=utf-8"))
-                                .withBody(responseBodyPage1)
+                                .withBody(responseBody)
                 );
 
-        mockServerClient.when(
-                request()
-                        .withMethod("GET")
-                        .withPath("/rest/api/1.0/projects/PAGED/repos")
-                        .withQueryStringParameter("start", "3")
-        )
-                .respond(
-                        response()
-                                .withStatusCode(200)
-                                .withHeader(header("Content-Type", "application/json; charset=utf-8"))
-                                .withBody(responseBodyPage2)
-                );
-
-        String projectKey = "PAGED";
-        repositories = client.getRepositories(projectKey);
+        String projectKey = "PRJ";
+        String repositorySlug = "REPO";
+        commits = client.getCommits(projectKey, repositorySlug);
     }
 
     @Test
-    public void countRepos() {
-        assertEquals("5 values in total", 5, repositories.size());
+    public void countCommits() {
+        assertEquals("1 commit in total", 1, commits.size());
     }
 
+    @Test
+    public void checkData() {
+        Commit commit = commits.iterator().next();
+        assertEquals("check commit id", "def0123abcdef4567abcdef8987abcdef6543abc", commit.getId());
+        assertEquals("check commit message", "More work on feature 1", commit.getMessage());
+    }
 }
