+++ src/test/java/at/grisa/agilemetrics/producer/bitbucketserver/restclient/BitBucketServerRestClientPagedEntitiesTest.java
@@ -1,7 +1,8 @@
-package at.grisa.agilemetrics.producer.bitbucketserver;
+package at.grisa.agilemetrics.producer.bitbucketserver.restclient;
 
 import at.grisa.agilemetrics.ApplicationConfig;
-import at.grisa.agilemetrics.producer.bitbucketserver.restentity.Project;
+import at.grisa.agilemetrics.producer.bitbucketserver.BitBucketServerRestClient;
+import at.grisa.agilemetrics.producer.bitbucketserver.restentity.Repository;
 import at.grisa.agilemetrics.util.CredentialManager;
 import org.junit.Before;
 import org.junit.Rule;
@@ -28,44 +29,53 @@ import static org.mockserver.model.HttpResponse.response;
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(classes = {BitBucketServerRestClient.class, CredentialManager.class, ApplicationConfig.class})
 @TestPropertySource("classpath:bitbucket-test.properties")
-public class BitBucketServerRestClientProjectsTest {
+public class BitBucketServerRestClientPagedEntitiesTest {
     @Autowired
     private BitBucketServerRestClient client;
 
     @Rule
     public MockServerRule mockServerRule = new MockServerRule(this, 1080);
     private MockServerClient mockServerClient;
-    private Collection<Project> projects;
+    private Collection<Repository> repositories;
 
     @Before
-    public void loadProjectsFromMockServer() throws IOException, URISyntaxException {
-        String responseBody = new String(Files.readAllBytes(Paths.get(getClass().getClassLoader().getResource("bitbucket/projects.js").toURI())));
+    public void loadReposFromMockServer() throws URISyntaxException, IOException {
+        String responseBodyPage1 = new String(Files.readAllBytes(Paths.get(getClass().getClassLoader().getResource("bitbucket/paged-entities-1.js").toURI())));
+        String responseBodyPage2 = new String(Files.readAllBytes(Paths.get(getClass().getClassLoader().getResource("bitbucket/paged-entities-2.js").toURI())));
 
         mockServerClient.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/rest/api/1.0/projects")
+                        .withPath("/rest/api/1.0/projects/PAGED/repos")
+                        .withQueryStringParameter("start", "0")
         )
                 .respond(
                         response()
                                 .withStatusCode(200)
                                 .withHeader(header("Content-Type", "application/json; charset=utf-8"))
-                                .withBody(responseBody)
+                                .withBody(responseBodyPage1)
                 );
 
-        projects = client.getProjects();
+        mockServerClient.when(
+                request()
+                        .withMethod("GET")
+                        .withPath("/rest/api/1.0/projects/PAGED/repos")
+                        .withQueryStringParameter("start", "3")
+        )
+                .respond(
+                        response()
+                                .withStatusCode(200)
+                                .withHeader(header("Content-Type", "application/json; charset=utf-8"))
+                                .withBody(responseBodyPage2)
+                );
+
+        String projectKey = "PAGED";
+        repositories = client.getRepositories(projectKey);
     }
 
     @Test
     public void countRepos() {
-        assertEquals("1 project in total", 1, projects.size());
+        assertEquals("5 values in total", 5, repositories.size());
     }
 
-    @Test
-    public void checkData() {
-        Project project = projects.iterator().next();
-        assertEquals("check project id", new Long(1), project.getId());
-        assertEquals("check project key", "PRJ", project.getKey());
-        assertEquals("check project name", "My Cool Project", project.getName());
-    }
 }
