+++ src/test/java/at/grisa/agilemetrics/producer/sonarqube/restclient/SonarQubeRestClientMeasuresTest.java
@@ -1,7 +1,9 @@
-package at.grisa.agilemetrics.producer.bitbucketserver;
+package at.grisa.agilemetrics.producer.sonarqube.restclient;
 
 import at.grisa.agilemetrics.ApplicationConfig;
-import at.grisa.agilemetrics.producer.bitbucketserver.restentity.Commit;
+import at.grisa.agilemetrics.producer.sonarqube.SonarQubeRestClient;
+import at.grisa.agilemetrics.producer.sonarqube.restentity.Measure;
+import at.grisa.agilemetrics.producer.sonarqube.restentity.Metric;
 import at.grisa.agilemetrics.util.CredentialManager;
 import org.junit.Before;
 import org.junit.Rule;
@@ -19,7 +21,6 @@ import java.net.URISyntaxException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.Collection;
-import java.util.Date;
 
 import static org.junit.Assert.assertEquals;
 import static org.mockserver.model.Header.header;
@@ -27,25 +28,27 @@ import static org.mockserver.model.HttpRequest.request;
 import static org.mockserver.model.HttpResponse.response;
 
 @RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes = {BitBucketServerRestClient.class, CredentialManager.class, ApplicationConfig.class})
-@TestPropertySource("classpath:bitbucket-test.properties")
-public class BitBucketServerRestClientCommitsFromTest {
+@ContextConfiguration(classes = {SonarQubeRestClient.class, CredentialManager.class, ApplicationConfig.class})
+@TestPropertySource("classpath:sonarqube-test.properties")
+public class SonarQubeRestClientMeasuresTest {
     @Autowired
-    private BitBucketServerRestClient client;
+    private SonarQubeRestClient client;
 
     @Rule
     public MockServerRule mockServerRule = new MockServerRule(this, 1080);
     private MockServerClient mockServerClient;
-    private Collection<Commit> commits;
+    private Collection<Measure> measures;
 
     @Before
-    public void loadCommitsFromMockServer() throws URISyntaxException, IOException {
-        String responseBody = new String(Files.readAllBytes(Paths.get(getClass().getClassLoader().getResource("bitbucket/commits-from.js").toURI())));
+    public void loadComponentFromMockServer() throws URISyntaxException, IOException {
+        String responseBody = new String(Files.readAllBytes(Paths.get(getClass().getClassLoader().getResource("sonarqube/component.js").toURI())));
 
         mockServerClient.when(
                 request()
                         .withMethod("GET")
-                        .withPath("/rest/api/1.0/projects/PRJ/repos/FROM/commits")
+                        .withPath("/api/measures/component")
+                        .withQueryStringParameter("component", "component")
+                        .withQueryStringParameter("metricKeys", "coverage")
         )
                 .respond(
                         response()
@@ -54,15 +57,18 @@ public class BitBucketServerRestClientCommitsFromTest {
                                 .withBody(responseBody)
                 );
 
-        String projectKey = "PRJ";
-        String repositorySlug = "FROM";
-        Date from = new Date();
-        from.setTime(1518587500000L);
-        commits = client.getCommits(projectKey, repositorySlug, from);
+        measures = client.getMeasures("component", Metric.COVERAGE);
     }
 
     @Test
-    public void countCommits() {
-        assertEquals("3 commits from Wed Feb 14 2018 05:51:40", 3, commits.size());
+    public void countIssues() {
+        assertEquals("1 measure in total", 1, measures.size());
+    }
+
+    @Test
+    public void checkData() {
+        Measure measure = measures.iterator().next();
+        assertEquals("check measure metric", "coverage", measure.getMetric());
+        assertEquals("check measure value", new Double(95.3), measure.getValue());
     }
 }
